refactor(versioning): Separate generic from loose versioning
