+++ lib/versioning/aws-machine-image/index.ts
@@ -1,5 +1,5 @@
 import { regEx } from '../../util/regex';
-import { GenericVersion, GenericVersioningApi } from '../loose/generic';
+import { GenericVersion, GenericVersioningApi } from '../generic';
 import type { VersioningApi } from '../types';
 
 export const id = 'aws-machine-image';
